refactor: refactored namespaces
