+++ src/Module.FeaturesManagementDashboard.Application/Commands/ICommandHandler.cs
@@ -1,6 +1,6 @@
 ﻿using MediatR;
 
-namespace Module.FeaturesManagementDashboard.Application.Commands
+namespace FeaturesManagementDashboard.Application.Commands
 {
     public interface ICommandHandler<in TCommand> : INotificationHandler<TCommand>
         where TCommand : class, ICommand
