+++ src/Module.FeaturesManagementDashboard.Application/Commands/UpdateFeature.cs
@@ -1,4 +1,4 @@
-﻿namespace Module.FeaturesManagementDashboard.Application.Commands
+﻿namespace FeaturesManagementDashboard.Application.Commands
 {
     public record UpdateFeature(string FeatureId, bool Status) : ICommand;
 }
