+++ src/Module.FeaturesManagementDashboard.Application/DI/ICompositionRoot.cs
@@ -1,6 +1,6 @@
 ﻿using SharedAbstractions.DI;
 
-namespace Module.FeaturesManagementDashboard.Application.DI
+namespace FeaturesManagementDashboard.Application.DI
 {
     public interface ICompositionRoot : IDependencyResolver
     {
