+++ src/Module.FeaturesManagementDashboard.Application/Queries/GetFeature.cs
@@ -1,6 +1,6 @@
-﻿using Module.FeaturesManagementDashboard.Application.DTO.Features;
+﻿using FeaturesManagementDashboard.Application.DTO.Features;
 
-namespace Module.FeaturesManagementDashboard.Application.Queries
+namespace FeaturesManagementDashboard.Application.Queries
 {
     public record GetFeature(string FeatureId) : IQuery<GetFeature, FeatureDto>;
 }
