+++ src/Module.FeaturesManagementDashboard.Application/Queries/IQuery.cs
@@ -1,6 +1,6 @@
 ﻿using MediatR;
 
-namespace Module.FeaturesManagementDashboard.Application.Queries
+namespace FeaturesManagementDashboard.Application.Queries
 {
     public interface IQuery<in TQuery, out TResult> : IRequest<TResult>
         where TQuery : class, IQuery<TQuery, TResult>
