+++ src/Module.FeaturesManagementDashboard.Application/Queries/IQueryHandler.cs
@@ -1,6 +1,6 @@
 ﻿using MediatR;
 
-namespace Module.FeaturesManagementDashboard.Application.Queries
+namespace FeaturesManagementDashboard.Application.Queries
 {
     public interface IQueryHandler<in TQuery, TResult> : IRequestHandler<TQuery, TResult>
         where TQuery : class, IRequest<TResult>
