+++ src/Module.FeaturesManagementDashboard.Domain/Entities/Features/FeatureId.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using SharedAbstractions.ValueObjects;
 
-namespace Module.FeaturesManagementDashboard.Domain.Entities.Features
+namespace FeaturesManagementDashboard.Domain.Entities.Features
 {
     public class FeatureId : ValueObject<string>
     {
