+++ src/Module.FeaturesManagementDashboard.Domain/Exceptions/DomainException.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace Module.FeaturesManagementDashboard.Domain.Exceptions
+namespace FeaturesManagementDashboard.Domain.Exceptions
 {
     public abstract class DomainException : Exception
     {
