+++ src/Module.FeaturesManagementDashboard.Domain/FeaturesManagementDashboard.Domain.csproj
