+++ src/Module.FeaturesManagementDashboard.Infrastructure/Dashboards/FeatureManagementDashboard.cs
@@ -2,7 +2,7 @@
 using Umbraco.Cms.Core.Composing;
 using Umbraco.Cms.Core.Dashboards;
 
-namespace Module.FeaturesManagementDashboard.Infrastructure.Dashboards
+namespace FeaturesManagementDashboard.Infrastructure.Dashboards
 {
     [Weight(80)]
     public class FeatureManagementDashboard : IDashboard
