+++ src/Module.FeaturesManagementDashboard.Infrastructure/Extensions/OptionsExtensions.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.Extensions.Configuration;
 
-namespace Module.FeaturesManagementDashboard.Infrastructure.Extensions
+namespace FeaturesManagementDashboard.Infrastructure.Extensions
 {
     internal static class OptionsExtensions
     {
