+++ src/Module.FeaturesManagementDashboard.Infrastructure/Extensions/SettingsExtensions.cs
@@ -1,15 +1,15 @@
-﻿using Microsoft.Extensions.Configuration;
+﻿using FeaturesManagementDashboard.Infrastructure.Settings;
+using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using Module.FeaturesManagementDashboard.Infrastructure.Settings;
 
-namespace Module.FeaturesManagementDashboard.Infrastructure.Extensions
+namespace FeaturesManagementDashboard.Infrastructure.Extensions
 {
 #pragma warning disable CS8604 // Possible null reference argument.
 
     internal static class SettingsExtensions
     {
         public static IServiceCollection AddSettings(this IServiceCollection serviceCollection)
-            => serviceCollection.AddSingleton<FeaturesManagementDashboardSettings>(factory
+            => serviceCollection.AddSingleton(factory
                 => factory.GetService<IConfiguration>()
                         .GetFeaturesManagementDashboardSettings());
