+++ src/Module.FeaturesManagementDashboard.Infrastructure/Queries/Handlers/GetFeatureHandler.cs
@@ -1,12 +1,12 @@
 ﻿using System.Threading;
 using System.Threading.Tasks;
-using Module.FeaturesManagementDashboard.Application.DTO.Features;
-using Module.FeaturesManagementDashboard.Application.Queries;
-using Module.FeaturesManagementDashboard.Domain.Entities.Features;
-using Module.FeaturesManagementDashboard.Domain.Repositories;
-using Module.FeaturesManagementDashboard.Infrastructure.Mappers;
+using FeaturesManagementDashboard.Application.DTO.Features;
+using FeaturesManagementDashboard.Application.Queries;
+using FeaturesManagementDashboard.Domain.Entities.Features;
+using FeaturesManagementDashboard.Domain.Repositories;
+using FeaturesManagementDashboard.Infrastructure.Mappers;
 
-namespace Module.FeaturesManagementDashboard.Infrastructure.Queries.Handlers
+namespace FeaturesManagementDashboard.Infrastructure.Queries.Handlers
 {
     internal class GetFeatureHandler : IQueryHandler<GetFeature, FeatureDto>
     {
